imports BuildInfoPlugin with related settings
workaround in case https://github.com/sbt/sbt-buildinfo/pull/55 is closed
