+++ src/main/scala/com/socrata/sbtplugins/CoreSettingsPlugin.scala
@@ -2,12 +2,9 @@ package com.socrata.sbtplugins
 
 import com.socrata.sbtplugins.CoveragePlugin.CoverageKeys.coverageDisable
 import com.socrata.sbtplugins.StylePlugin.StyleKeys.styleCheck
-import org.joda.time.{DateTime, DateTimeZone}
 import sbt.Keys._
 import sbt._
 import sbtassembly.AssemblyKeys.assembly
-import sbtbuildinfo.BuildInfoPlugin
-import sbtbuildinfo.BuildInfoPlugin.autoImport._
 
 /** Enables autoplugins, sets scala version and compiler flags for static analysis. */
 object CoreSettingsPlugin extends AutoPlugin {
@@ -15,8 +12,8 @@ object CoreSettingsPlugin extends AutoPlugin {
     * @return On all requirements in all scopes. */
   override def trigger: PluginTrigger = allRequirements
   /** Depends on these autoplugins.
-    * @return Basic jvm, style, coverage. */
-  override def requires: Plugins = plugins.JvmPlugin && BuildInfoPlugin
+    * @return Basic jvm, others enabled by trigger. */
+  override def requires: Plugins = plugins.JvmPlugin
 
   private val compileEncoding = Seq("-encoding", "UTF-8")
   private val compileDebug29 = Seq("-g:vars")
@@ -28,15 +25,6 @@ object CoreSettingsPlugin extends AutoPlugin {
   override def projectSettings: Seq[Def.Setting[_]] = Seq(
     test in assembly := {},
     assembly <<= assembly dependsOn (styleCheck in Compile, coverageDisable),
-    buildInfoKeys ++= Seq[BuildInfoKey](
-      name,
-      version,
-      scalaVersion,
-      sbtVersion,
-      BuildInfoKey.action("buildTime") { DateTime.now(DateTimeZone.UTC).toDateTimeISO },
-      BuildInfoKey.action("revision") { gitRevision }
-    ),
-    buildInfoOptions ++= Seq(BuildInfoOption.ToMap, BuildInfoOption.ToJson),
     scalaVersion := "2.10.5",
     scalacOptions ++= compileEncoding ++ Seq("-Xlint", "-deprecation", "-Xfatal-warnings", "-unchecked"),
     scalacOptions <++= scalaVersion map {
@@ -67,23 +55,6 @@ object CoreSettingsPlugin extends AutoPlugin {
 
   object SocrataSbtKeys {
     val dependenciesSnippet = SettingKey[xml.NodeSeq]("socrata-dependencies-snippet")
-  }
-
-  private class SimpleProcessLog extends ProcessLogger {
-    var errorString: Option[String] = None
-    var infoString: Option[String] = None
-    override def buffer[T](f: => T): T = f
-    override def error(s: => String): Unit = errorString = Some(s)
-    override def info(s: => String): Unit = infoString = Some(s)
-  }
-
-  def gitRevision: String = {
-    val procLog = new SimpleProcessLog
-    val exitCode = Process(Seq("git", "describe", "--always", "--dirty", "--long", "--abbrev=40")).!(procLog)
-    exitCode match {
-      case 0 => procLog.infoString.getOrElse("git stdout = null")
-      case n: Int => s"git error $n: ${procLog.errorString.getOrElse("git stderr = null")}"
-    }
   }
 
   object ScalaVersion {
