+++ src/test/scala/com/socrata/sbtplugins/BuildInfoSpec.scala
@@ -3,6 +3,12 @@ package com.socrata.sbtplugins
 import org.scalatest.{FunSuiteLike, Matchers}
 
 class BuildInfoSpec extends FunSuiteLike with Matchers {
+  test("BuildInfoPlugin is an autoplugin") {
+    Option(BuildInfoPlugin.requires) should be('defined)
+    Option(BuildInfoPlugin.trigger) should be('defined)
+    Option(BuildInfoPlugin.projectSettings) should be('defined)
+  }
+
   test("BuildInfo includes these fields") {
     Option(BuildInfo.name) should be('defined)
     Option(BuildInfo.version) should be('defined)
@@ -14,5 +20,9 @@ class BuildInfoSpec extends FunSuiteLike with Matchers {
 
   test("BuildInfo json") {
     Option(BuildInfo.toJson) should be('defined)
+  }
+
+  test("BuildInfoPlugin defines git revision") {
+    Option(BuildInfoPlugin.gitRevision) should be('defined)
   }
 }
